move save state tracking from middleware/data -> store/saveState. ignore /screenshots
