+++ src/middleware/data.js
@@ -3,48 +3,7 @@ import invariant from 'invariant';
 import { headersGet, headersPost, headersPut, headersDelete } from './headers';
 import { dataApiPath } from './paths';
 import * as instanceMap from '../store/instanceMap';
-
-/******
- save state
- ******/
-
-//track information about saving e.g. time
-const saveState = new Map();
-
-const noteSave = (projectId, sha = null) => {
-  invariant(projectId, 'must pass project ID');
-  const lastState = saveState.get(projectId) || {};
-
-  saveState.set(projectId, Object.assign(lastState, {
-    lastSaved: +Date.now(),
-    sha,
-  }));
-};
-
-const noteFailure = (projectId, err) => {
-  invariant(projectId, 'must pass project ID');
-  const lastState = saveState.get(projectId) || {};
-
-  saveState.set(projectId, Object.assign(lastState, {
-    lastFailed: +Date.now(),
-    lastErr: err,
-  }));
-};
-
-export const getProjectSaveState = (projectId) => {
-  invariant(projectId, 'must pass project ID');
-  const state = saveState.get(projectId) || {};
-  const { lastSaved = 0, lastFailed = 0, sha = null, lastErr = null } = state;
-
-  return {
-    lastSaved,
-    sha,
-    lastFailed,
-    lastErr,
-    saveDelta: +Date.now() - lastSaved,
-    saveSuccessful: lastFailed <= lastSaved,
-  };
-};
+import { noteSave, noteFailure } from '../store/saveState';
 
 /******
  API requests
